+++ pkgs/test_api/lib/src/backend/declarer.dart
@@ -91,6 +91,12 @@ class Declarer {
   /// Whether any tests and/or groups have been flagged as solo.
   bool get _solo => _soloEntries.isNotEmpty;
 
+  /// A full test name to match when running in single test mode.
+  ///
+  /// Groups which are not a strict prefix of `filterToSingleTestName` will be
+  /// skipped.
+  final String? _filterToSingleTestName;
+
   /// The current zone-scoped declarer.
   static Declarer? get current => Zone.current[#test.declarer] as Declarer?;
 
@@ -113,7 +119,8 @@ class Declarer {
       {Metadata? metadata,
       Set<String>? platformVariables,
       bool collectTraces = false,
-      bool noRetry = false})
+      bool noRetry = false,
+      String? filterToSingleTestName})
       : this._(
             null,
             null,
@@ -121,10 +128,19 @@ class Declarer {
             platformVariables ?? const UnmodifiableSetView.empty(),
             collectTraces,
             null,
-            noRetry);
+            noRetry,
+            filterToSingleTestName);
 
-  Declarer._(this._parent, this._name, this._metadata, this._platformVariables,
-      this._collectTraces, this._trace, this._noRetry);
+  Declarer._(
+    this._parent,
+    this._name,
+    this._metadata,
+    this._platformVariables,
+    this._collectTraces,
+    this._trace,
+    this._noRetry,
+    this._filterToSingleTestName,
+  );
 
   /// Runs [body] with this declarer as [Declarer.current].
   ///
@@ -143,6 +159,12 @@ class Declarer {
       bool solo = false}) {
     _checkNotBuilt('test');
 
+    final fullName = _prefix(name);
+    if (_filterToSingleTestName != null &&
+        fullName != _filterToSingleTestName) {
+      return;
+    }
+
     var newMetadata = Metadata.parse(
         testOn: testOn,
         timeout: timeout,
@@ -152,8 +174,7 @@ class Declarer {
         retry: _noRetry ? 0 : retry);
     newMetadata.validatePlatformSelectors(_platformVariables);
     var metadata = _metadata.merge(newMetadata);
-
-    _entries.add(LocalTest(_prefix(name), metadata, () async {
+    _entries.add(LocalTest(fullName, metadata, () async {
       var parents = <Declarer>[];
       for (Declarer? declarer = this;
           declarer != null;
@@ -195,6 +216,11 @@ class Declarer {
       bool solo = false}) {
     _checkNotBuilt('group');
 
+    final fullName = _prefix(name);
+    if (!(_filterToSingleTestName?.startsWith(fullName) ?? true)) {
+      return;
+    }
+
     var newMetadata = Metadata.parse(
         testOn: testOn,
         timeout: timeout,
@@ -207,7 +233,7 @@ class Declarer {
     var trace = _collectTraces ? Trace.current(2) : null;
 
     var declarer = Declarer._(this, _prefix(name), metadata, _platformVariables,
-        _collectTraces, trace, _noRetry);
+        _collectTraces, trace, _noRetry, _filterToSingleTestName);
     declarer.declare(() {
       // Cast to dynamic to avoid the analyzer complaining about us using the
       // result of a void method.