+++ pkgs/test_api/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.19-nullsafety.3
+
+* Add capability for ensuring unique test names and collecting all test names to
+  `Declarer.`
+* Add capability to filter to a single unique test name in `Declarer`.
+
 ## 0.2.19-nullsafety.2
 
 * Allow `2.10` stable and `2.11.0-dev` SDKs.